\name{curse-package}
\alias{curse-package}
\docType{package}
\title{
\packageTitle{curse}
}
\description{
\packageDescription{curse}
}
\details{

Package: curse
Type: Package
Version: 1.0
Date: 2021-03-02
License: MIT + file LICENSE

\packageDESCRIPTION{curse}
\packageIndices{curse}

}
\author{
\packageAuthor{curse}

Maintainer: \packageMaintainer{curse}
}
\keyword{ package }
