% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curse.R
\name{curse}
\alias{curse}
\title{Random cursing words/phrases}
\usage{
curse(polite = FALSE)
}
\arguments{
\item{polite}{Logical. If FALSE (default), returns the one of the original cursing
phrases. If TRUE, the 'eff' word is blanked by three stars}
}
\value{
A randomly sampled cursing word or phrase
}
\description{
This function prints cursing words on the screen. The user can choose between
the original content or a 'polite' version.
}
\examples{
curse(polite=TRUE)
}
