\name{phrases}
\alias{phrases}
\docType{data}
\title{
List of cursing words and phrases
}
\description{
 List of cursing words and phrases
}
\usage{data("phrases")}
\format{
  A dataset with 119 cursing words and phrases.
  \describe{
  \item{phrases}{List of cursing words and phrases}
  }
}
\source{
}
\examples{
}
\keyword{datasets}
