\name{cuRse-package}
\alias{cuRse-package}
\docType{package}
\title{
\packageTitle{cuRse}
}
\description{
\packageDescription{cuRse}
}
\details{

Package: cuRse
Type: Package
Version: 1.0
Date: 2021-03-02
License: MIT + file LICENSE

\packageDESCRIPTION{cuRse}
\packageIndices{cuRse}

}
\author{
\packageAuthor{cuRse}

Maintainer: \packageMaintainer{cuRse}
}
\keyword{ package }
