% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curse-package.R, R/curse.R
\docType{package}
\name{curse}
\alias{curse}
\alias{curse-package}
\title{curse package}
\usage{
curse()
}
\description{
Express anger or frustration in code

Simple memory-less cursing function
}
\examples{
curse()
}
\author{
\strong{Maintainer}: Alessandro Gentile Last \email{alessandro.gentile03@gmail.com}

}
